# backend/API/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SipScore API",
    "version": "1.0.0",
    "description": "API documentation for SipScore application"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users",
        "description": "Returns all verified, non-deleted users.",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "Signup",
        "description": "Create a new user and send verification code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created, verification sent" },
          "400": { "description": "Invalid email format" },
          "409": { "description": "Email or username already in use" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by ID",
        "description": "Returns a single verified, non-deleted user by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Update User / Verify Email",
        "description": "Update user details or verify email with code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "request": { "type": "object" },
                  "action": { "type": "string" }
                },
                "required": ["id", "code"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated / Email verified" },
          "400": { "description": "You fucked up: {error}" }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete user",
        "description": "Soft delete user by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "User deleted" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/roles/{id}": {
      "put": {
        "tags": ["Role"],
        "summary": "Promote user role",
        "description": "Change the role of a user (admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newRole": { "type": "integer", "minimum": 0, "maximum": 3 }
                },
                "required": ["newRole"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User role updated" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "User not found" },
          "400": { "description": "Invalid role specified" }
        }
      }
    },
...
  }
}